requirements.txt
absl-py==2.1.0
aiofiles==24.1.0
aiogram==3.14.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiopytesseract==0.14.0
aiosignal==1.3.1
annotated-types==0.7.0
astunparse==1.6.3
attrs==24.2.0
Brotli==1.1.0
cattrs==24.1.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cloudscraper==1.2.71
contourpy==1.3.0
curl_cffi==0.7.3
cycler==0.12.1
editdistance==0.8.1
efficientnet==1.0.0
essential-generators==1.0
flatbuffers==24.3.25
fonttools==4.54.1
frozenlist==1.5.0
g4f==0.3.3.3
gast==0.6.0
google-pasta==0.2.0
grpcio==1.67.1
h5py==3.12.1
idna==3.10
imageio==2.36.0
imgaug==0.4.0
keras==3.6.0
Keras-Applications==1.0.8
kiwisolver==1.4.7
lazy_loader==0.4
libclang==18.1.1
magic-filter==1.0.12
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
mdurl==0.1.2
ml_dtypes==0.5.0
multidict==6.1.0
namex==0.0.8
networkx==3.4.2
numpy==2.0.2
opencv-python==4.10.0.84
opt_einsum==3.4.0
optree==0.13.0
packaging==24.2
pillow==11.0.0
propcache==0.2.0
protobuf==5.28.3
pyclipper==1.3.0.post6
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
pyparsing==3.2.0
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
scikit-image==0.24.0
scipy==1.14.1
setuptools==75.3.0
shapely==2.0.6
six==1.16.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
termcolor==2.5.0
tesseract==0.1.3
tifffile==2024.9.20
tqdm==4.67.0
typing_extensions==4.12.2
urllib3==2.2.3
validators==0.34.0
Werkzeug==3.1.3
wheel==0.45.0
wrapt==1.16.0
yarl==1.17.1
